+++ notification/core/src/main/java/de/taimos/dvalin/notification/template/ITemplateResolver.java
@@ -1,0 +1,11 @@
+package de.taimos.dvalin.notification.template;
+
+import java.util.Map;
+
+public interface ITemplateResolver {
+
+    String resolveTemplate(String location, Map<String, Object> context);
+
+    String resolveRawTemplate(String template, Map<String, Object> context);
+
+}